Extract AndroidTeamGfiSummary and TeamLeadGfiDonutSection into separate component files
