+++ features/dashboard/views/team-lead-view/tabs/overview.tab.tsx
@@ -4,11 +4,8 @@ import { useEffect, useState } from "react";
 import {
   Area,
   CartesianGrid,
-  Cell,
   Line,
   LineChart,
-  Pie,
-  PieChart,
   ResponsiveContainer,
   Tooltip,
   XAxis,
@@ -16,7 +13,6 @@ import {
 } from "recharts";
 import {
   AlertCircle,
-  AlertTriangle,
   ArrowUpRight,
   ChartPie,
   LayoutDashboard,
@@ -28,11 +24,6 @@ import {
 import { getRoleDisplayLabel } from "@/lib/auth/role-display";
 import { formatDisplayValue } from "@/lib/utils/display.utils";
 import { Skeleton } from "@/components/ui/skeleton";
-import {
-  Tooltip as UiTooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
 import type {
   TeamReport,
   TeamLeadOverviewResponse,
@@ -40,7 +31,6 @@ import type {
 } from "../overview.types";
 import {
   CHART_GRADIENT_ID,
-  GFI_DOMAIN_COLORS,
   STAT_CARD_STYLES,
   formatChartTickLabel,
   formatChartTooltipLabel,
@@ -50,185 +40,8 @@ import {
   getTeamLeadActionMessage,
   getTotalGfiCount,
 } from "../overview.utils";
-
-function AndroidTeamGfiSummary({
-  counts,
-}: {
-  counts: TeamReport["gfiCounts"];
-}) {
-  const total = getTotalGfiCount(counts);
-  const progress = Math.min((total / 15) * 100, 100);
-  const statusLabel =
-    total >= 15
-      ? "Healthy pipeline"
-      : total >= 8
-        ? "Needs growth"
-        : "Low stock";
-
-  return (
-    <div className="relative overflow-hidden rounded-xl border border-slate-200 bg-[radial-gradient(circle_at_top_left,_rgba(148,163,184,0.14),_transparent_40%),linear-gradient(135deg,_#ffffff_0%,_#f8fafc_55%,_#f1f5f9_100%)] px-5 py-5">
-      <div className="pointer-events-none absolute -right-10 -top-10 h-28 w-28 rounded-full bg-slate-200/50 blur-2xl" />
-      <div className="pointer-events-none absolute -bottom-8 left-6 h-20 w-20 rounded-full bg-sky-100/50 blur-2xl" />
-
-      <div className="relative flex items-start justify-between gap-4">
-        <div>
-          <p className="text-xs font-semibold uppercase tracking-[0.16em] text-slate-500">
-            Android GFI Pool
-          </p>
-          <p className="mt-3 text-4xl font-semibold leading-none text-slate-950">
-            {total}
-          </p>
-          <p className="mt-2 text-sm text-slate-600">
-            Open starter issues available for Android contributors.
-          </p>
-        </div>
-
-        <span className="rounded-full border border-white/80 bg-white/75 px-3 py-1 text-xs font-medium text-slate-700 shadow-sm backdrop-blur">
-          {statusLabel}
-        </span>
-      </div>
-
-      <div className="relative mt-5 rounded-xl border border-white/70 bg-white/80 p-4 backdrop-blur-sm">
-        <div className="flex items-center justify-between gap-3">
-          <p className="text-xs font-semibold uppercase tracking-[0.14em] text-slate-500">
-            Coverage Progress
-          </p>
-          <p className="text-sm font-medium text-slate-700">
-            {progress.toFixed(0)}%
-          </p>
-        </div>
-        <div className="mt-3 h-2.5 overflow-hidden rounded-full bg-slate-200/80">
-          <div
-            className="h-full rounded-full bg-[linear-gradient(90deg,_#334155_0%,_#0f172a_100%)] transition-all"
-            style={{ width: `${progress}%` }}
-          />
-        </div>
-        <p className="mt-3 text-sm text-slate-600">
-          Targeting a steady pool of 15+ open good first issues keeps Android
-          onboarding simple and visible.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-function TeamLeadGfiDonutSection({
-  counts,
-}: {
-  counts: TeamReport["gfiCounts"];
-}) {
-  const data = [
-    {
-      color: GFI_DOMAIN_COLORS.frontend,
-      key: "frontend",
-      label: "Frontend",
-      value: counts.frontend,
-    },
-    {
-      color: GFI_DOMAIN_COLORS.backend,
-      key: "backend",
-      label: "Backend",
-      value: counts.backend,
-    },
-    {
-      color: GFI_DOMAIN_COLORS.fullstack,
-      key: "fullstack",
-      label: "Fullstack",
-      value: counts.fullstack,
-    },
-    {
-      color: GFI_DOMAIN_COLORS.uncategorized,
-      key: "uncategorized",
-      label: "Uncategorized",
-      value: counts.uncategorized,
-    },
-  ] as const;
-  const total = data.reduce((sum, item) => sum + item.value, 0);
-  const chartData =
-    total === 0
-      ? [{ color: "#e2e8f0", key: "empty", label: "No GFIs", value: 1 }]
-      : data;
-
-  return (
-    <div className="flex flex-col items-center gap-5 sm:flex-row sm:items-center">
-      <div className="mx-auto h-52 w-full max-w-[220px] shrink-0">
-        <ResponsiveContainer width="100%" height="100%">
-          <PieChart>
-            <Pie
-              data={chartData}
-              dataKey="value"
-              innerRadius={58}
-              outerRadius={84}
-              paddingAngle={total === 0 ? 0 : 3}
-              strokeWidth={0}
-            >
-              {chartData.map((item) => (
-                <Cell key={item.key} fill={item.color} />
-              ))}
-            </Pie>
-            <text
-              x="50%"
-              y="47%"
-              textAnchor="middle"
-              className="fill-slate-400 text-[11px] uppercase tracking-[0.16em]"
-            >
-              Total GFIs
-            </text>
-            <text
-              x="50%"
-              y="58%"
-              textAnchor="middle"
-              className="fill-slate-900 text-[28px] font-semibold"
-            >
-              {total}
-            </text>
-          </PieChart>
-        </ResponsiveContainer>
-      </div>
-
-      <div className="flex min-w-0 flex-1 flex-col gap-2">
-        {data.map((item) => (
-          <div
-            key={item.key}
-            className="rounded-xl border border-slate-200 bg-white px-4 py-3"
-          >
-            <div className="flex items-center justify-between gap-3">
-              <div className="flex items-center gap-2">
-                <span
-                  className="h-3.5 w-3.5 shrink-0 rounded-sm"
-                  style={{ backgroundColor: item.color }}
-                />
-                <p className="truncate text-sm font-medium text-slate-700">
-                  {item.label}
-                </p>
-                {item.key !== "uncategorized" && item.value < 5 && (
-                  <UiTooltip>
-                    <TooltipTrigger asChild>
-                      <span className="inline-flex">
-                        <AlertTriangle
-                          className="h-4 w-4 shrink-0 text-red-500"
-                          aria-label={`${item.label} has fewer than 5 good first issues`}
-                        />
-                      </span>
-                    </TooltipTrigger>
-                    <TooltipContent>
-                      <p>
-                        Each domain should have at least 5 good first issues.
-                      </p>
-                    </TooltipContent>
-                  </UiTooltip>
-                )}
-              </div>
-              <p className="shrink-0 text-lg font-semibold text-slate-900">
-                {item.value}
-              </p>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
+import { AndroidTeamGfiSummary } from "../components/android-team-gfi-summary";
+import { TeamLeadGfiDonutSection } from "../components/team-lead-gfi-donut-section";
 
 function TeamOverviewChartTooltip({
   active,
